requirements.txt
fastapi
sqlmodel
uvicorn
python-multipart
python-jose[cryptography]
python_nbt
anyio
pydantic